requirements.txt
pydantic>=2.7
pyyaml>=6.0.1
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.2
python-dotenv>=1.0
rich>=13.7
python-docx>=1.1
